girepository: Fix straightforward -Wsign-conversion warnings

Helps: #3405

+++ girepository/giarginfo.c
@@ -286,7 +286,7 @@ gi_arg_info_get_closure_index (GIArgInfo    *info,
   has_closure_index = (blob->closure >= 0);
 
   if (out_closure_index != NULL)
-    *out_closure_index = has_closure_index ? blob->closure : 0;
+    *out_closure_index = has_closure_index ? (unsigned int) blob->closure : 0;
   return has_closure_index;
 }
 
@@ -317,7 +317,7 @@ gi_arg_info_get_destroy_index (GIArgInfo    *info,
   has_destroy_index = (blob->destroy >= 0);
 
   if (out_destroy_index != NULL)
-    *out_destroy_index = has_destroy_index ? blob->destroy : 0;
+    *out_destroy_index = has_destroy_index ? (unsigned int) blob->destroy : 0;
   return has_destroy_index;
 }
